+++ component/azstorage/config.go
@@ -46,7 +46,7 @@ import (
 	"github.com/JeffreyRichter/enum/enum"
 )
 
-//  AuthType Enum
+// AuthType Enum
 type AuthType int
 
 var EAuthType = AuthType(0).INVALID_AUTH()
@@ -83,7 +83,7 @@ func (a *AuthType) Parse(s string) error {
 	return err
 }
 
-//  AccountType Enum
+// AccountType Enum
 type AccountType int
 
 var EAccountType = AccountType(0).INVALID_ACC()
@@ -177,7 +177,7 @@ type AzStorageOptions struct {
 	CaCertFile     string `config:"ca-cert-file" yaml:"-"`
 }
 
-//  RegisterEnvVariables : Register environment varilables
+// RegisterEnvVariables : Register environment varilables
 func RegisterEnvVariables() {
 	config.BindEnv("azstorage.account-name", EnvAzStorageAccount)
 	config.BindEnv("azstorage.type", EnvAzStorageAccountType)