+++ internal/handlemap/handle_map.go
@@ -153,7 +153,7 @@ func (handle *Handle) Cleanup() {
 	handle.Unlock()
 }
 
-//defaultHandleMap holds a synchronized map[ HandleID ]*Handle
+// defaultHandleMap holds a synchronized map[ HandleID ]*Handle
 var defaultHandleMap sync.Map
 var nextHandleID = *atomic.NewUint64(uint64(0))
 
@@ -199,8 +199,8 @@ func Load(key HandleID) (*Handle, bool) {
 	return handle, true
 }
 
-//Store function must not be used in production application.
-//This is a utility function present only for test scenarios.
+// Store function must not be used in production application.
+// This is a utility function present only for test scenarios.
 func Store(key HandleID, path string, _ uintptr) *Handle {
 	handle := &Handle{
 		ID:   key,