+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## 2.0.4 (WIP)
+**Features**
+- Added new config parameter "max-fuse-threads" under "libfuse" config to control max threads allowed at libfuse layer.
+
 **Bug Fixes**
 - [[#1116](https://github.com/Azure/azure-storage-fuse/issues/1116)] Relative path for tmp-cache is resulting into file read-write failure.